refacto v2 skibidi
